// mobile/src/types/themeColors.d.ts
export type ThemeColors = {
	dark: boolean
	colors: {
		text: string
		textMuted: string
		background: string
		tint: string
		tabIconDefault: string
		borderInput: string
		tabIconSelected: string
		ImputBackgroundColors: string
		colorIconInput: string
		secondaryMuted: string
		primary: string
		primaryMuted: string
		secondary: string
		card: string
		border: string
		notification: string
		muted: string
		tabBarBackgroundColor: string
		tabBarActiveTintColor: string
		modal: string
	}
	fonts: unknown
}